+++ apps/bare-expo/android/app/src/main/java/dev/expo/payments/MainActivity.java
@@ -6,12 +6,12 @@ import android.content.SharedPreferences;
 import android.net.Uri;
 import android.os.Bundle;
 
+import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
 
 import expo.modules.ReactActivityDelegateWrapper;
-import expo.modules.ReactActivityWrapper;
 
-public class MainActivity extends ReactActivityWrapper {
+public class MainActivity extends ReactActivity {
 
   /**
    * Returns the name of the main component registered from JavaScript.
