+++ packages/expo-dev-launcher/android/src/release/java/expo/modules/devlauncher/DevLauncherPackageDelegate.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import expo.modules.core.interfaces.ApplicationLifecycleListener
-import expo.modules.core.interfaces.ReactActivityDelegateHandler
 import expo.modules.core.interfaces.ReactActivityLifecycleListener
 import expo.modules.core.interfaces.ReactActivityHandler
 
@@ -12,5 +11,5 @@ object DevLauncherPackageDelegate {
   fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> = emptyList()
   fun createApplicationLifecycleListeners(context: Context?): List<ApplicationLifecycleListener> = emptyList()
   fun createReactActivityLifecycleListeners(activityContext: Context?): List<ReactActivityLifecycleListener> = emptyList()
-  fun createReactActivityDelegateHandlers(activityContext: Context?): List<ReactActivityDelegateHandler> = emptyList()
+  fun createReactActivityHandlers(activityContext: Context?): List<ReactActivityHandler> = emptyList()
 }
