+++ packages/expo-modules-core/android/src/main/java/expo/modules/core/BasePackage.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import expo.modules.core.interfaces.ApplicationLifecycleListener;
 import expo.modules.core.interfaces.InternalModule;
 import expo.modules.core.interfaces.Package;
-import expo.modules.core.interfaces.ReactActivityDelegateHandler;
 import expo.modules.core.interfaces.ReactActivityHandler;
 import expo.modules.core.interfaces.ReactActivityLifecycleListener;
 import expo.modules.core.interfaces.ReactNativeHostHandler;
@@ -49,11 +48,6 @@ public class BasePackage implements Package {
 
   @Override
   public List<ReactActivityLifecycleListener> createReactActivityLifecycleListeners(Context activityContext) {
-    return Collections.emptyList();
-  }
-
-  @Override
-  public List<ReactActivityDelegateHandler> createReactActivityDelegateHandlers(Context activityContext) {
     return Collections.emptyList();
   }
